修bug

+++ src/main/java/cn/tycoding/domain/Cargo.java
@@ -28,7 +28,7 @@ public class Cargo implements Serializable {
     private Integer preCargoId;
 
     // 赔偿金，是订单超时要进行赔偿的依据，也是对车辆担保额的要求
-    private int insurance;
+    private double insurance;
     // 当前挂单的运费
     private double freightFare;
     private int receiverId;
